Move asset-version and task-arguments files
