+++ tasks/gulp/compile-assets.js
@@ -11,7 +11,7 @@ const plumber = require('gulp-plumber')
 const postcss = require('gulp-postcss')
 const autoprefixer = require('autoprefixer')
 const rollup = require('gulp-better-rollup')
-const taskArguments = require('./task-arguments')
+const taskArguments = require('../task-arguments')
 const gulpif = require('gulp-if')
 const uglify = require('gulp-uglify')
 const eol = require('gulp-eol')
