+++ tasks/gulp/copy-to-destination.js
@@ -5,7 +5,7 @@ const gulp = require('gulp')
 const configPaths = require('../../config/paths.json')
 const postcss = require('gulp-postcss')
 const autoprefixer = require('autoprefixer')
-const taskArguments = require('./task-arguments')
+const taskArguments = require('../task-arguments')
 const filter = require('gulp-filter')
 const fs = require('fs')
 const yamlToJson = require('js-yaml')
